% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MAZE}
\alias{MAZE}
\title{Mediation Analysis for ZEro-inflated mediators}
\usage{
MAZE(
  data,
  distM = c("zilonm", "zinbm", "zipm"),
  X,
  M,
  Y,
  Z = NULL,
  x1,
  x2,
  zval = NULL,
  B = 20,
  seed = 1
)
}
\arguments{
\item{data}{a data frame containing variables an independent variable X, a mediator M, an outcome Y, and confounder variables Z (if any). See example dataset: data(zip10)}

\item{distM}{an optional character value for distribution to be used for the mediator. Possible choices are 'zilonm','zinbm', or 'zipm' for zero-inflated log-normal, negative binomial, or Poisson mediators respectively. By default, all three distributions will be fitted and the final mediation model is selected by AIC}

\item{X}{name of the independent variable}

\item{M}{name of the mediator variable}

\item{Y}{name of the outcome variable}

\item{Z}{name(s) of confounder variables}

\item{x1}{the first value of independent variable of interest}

\item{x2}{the second value of independent variable of interest}

\item{zval}{the value of confounders to be conditional on in estimating effects}

\item{B}{the upper bound value B to be used in the probability mechanism of observing false zeros}

\item{seed}{an optional seed number to control randomness}
}
\value{
a list containing:
\itemize{
\item results_effects: a data frame for the results of estimated mediation effect,
\item results_parameters: a data frame for the results of model parameters,
\item BIC: a numeric value for the BIC of the final mediation model,
\item AIC: a numeric value for the AIC of the final mediation model,
\item selected_distM: a character value for the distribution of M selected in the final mediation model
}
}
\description{
Mediation Analysis for ZEro-inflated mediators
}
\details{
A novel mediation modeling approach to address zero-inflated mediators containing true zeros and false zeros.

For an independent variable \eqn{X}, a zero-inflated mediator \eqn{M} and a continuous outcome variable \eqn{Y}, the following regression equation is used to model the association between $Y$ and $(X,M)$:
\deqn{Y_{xm1_{(m>0)}}=\beta_0+\beta_1m+\beta_2 1_{(m>0)}+\beta_3x+\beta_4x1_{(m>0)}+\beta_5xm+\epsilon}
}
\examples{
data(zip10)
\dontrun{
MAZE(data=zip10, distM=c('zilonm', 'zinbm', 'zipm'), 
X='X', M='Mobs', Y='Y', Z=NULL, x1=0, x2=1, B=20, seed=1)
}
}
\author{
Zhigang Li \href{mailto:zhigang.li@ufl.edu}{zhigang.li@ufl.edu}
Meilin Jiang \href{mailto:meilin.jiang@ufl.edu}{meilin.jiang@ufl.edu}
}
