% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot a maze's graph.}
\usage{
plot_graph(g, layout = NULL, labels = TRUE, edge.color = "black",
  vertex.color = "black", vertex.fill = "white", vertex.size = 15,
  vertex.label.color = "white", vertex.label.size = 1, path.show = FALSE,
  path.start = 1, path.end = vcount(g), ...)
}
\arguments{
\item{g}{graph with maze structure.}

\item{layout}{layout (default taken from g$layout).}

\item{labels}{logical; whether to plot node labels.}

\item{edge.color}{color used for edges.}

\item{vertex.color}{color used for node borders.}

\item{vertex.fill}{color used for nodes.}

\item{vertex.size}{node size.}

\item{vertex.label.color}{color used for node labels.}

\item{path.show}{logical; whether to show the path between two nodes.}

\item{path.start}{path starting node.}

\item{path.end}{path end node.}

\item{...}{extra arguments passed down to geom_segment().}
}
\description{
Plot a maze's graph.
}

