% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_maze}
\alias{plot_maze}
\title{Plot a simple maze.}
\usage{
plot_maze(g, wall.size = 5, tile.show = FALSE, tile.size = 1,
  tile.color = "white", tile.number.show = FALSE,
  tile.number.size = 5, path.show = FALSE, path.start = 1,
  path.end = vcount(g))
}
\arguments{
\item{g}{graph with maze structure.}

\item{wall.size}{size of walls.}

\item{tile.show}{logical; whether to show tile borders.}

\item{tile.size}{size of tile borders.}

\item{tile.color}{color of tile borders.}

\item{tile.number.show}{logical; whether to show tile number.}

\item{tile.number.size}{size of tile numbers.}

\item{path.show}{logical; whether to show the path between two tiles.}

\item{path.start}{starting point for path.}

\item{path.end}{end point for path.}
}
\description{
Plot a simple maze.
}
\examples{
NULL
}
