% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLmapper.R
\name{RLmapper}
\alias{RLmapper}
\title{Create RLmapper}
\usage{
RLmapper(qvalue, adjacency, num_interval = 10)
}
\arguments{
\item{qvalue}{Vector of state values}

\item{adjacency}{Adjacency matrix of states}

\item{num_interval}{Division number}
}
\value{
Adjacency matrix of set of states
}
\description{
Create RLmapper
}
