% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip10.R
\docType{data}
\name{zip10}
\alias{zip10}
\title{Example dataset 'zip10'}
\format{
A data frame with 300 rows and 3 variables:
\describe{
\item{X}{independent variable, continuous data type}
\item{Y}{outcome, contiuous data type}
\item{Mobs}{observed mediator values with possibly false zeros, count data type}
}
}
\usage{
zip10
}
\description{
An example dataset generated from the proposed model with a zero-inflated Poisson mediator. About 7\% of zero values are included
}
\keyword{datasets}
