% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MAZE}
\alias{MAZE}
\title{Mediation Analysis for ZEro-inflated mediators}
\usage{
MAZE(
  data,
  distM = c("zilonm", "zinbm", "zipm"),
  X,
  M,
  Y,
  x1,
  x2,
  B = 20,
  seed = 1
)
}
\arguments{
\item{data}{a data frame containing variables an independent variable, a mediator, and an outcome variables. See example dataset: data(zip10)}

\item{distM}{an optional character value for distribution to be used for the mediator. Possible choices are 'zilonm','zinbm', or 'zipm' for zero-inflated log-normal, negative binomial, or Poisson mediators respectively. By default, all three distributions will be fitted and the final mediation model is selected by AIC}

\item{X}{variable name of the independent variable}

\item{M}{variable name of the mediator variable}

\item{Y}{variable name of the outcome variable}

\item{x1}{the first value of independent variable of interest}

\item{x2}{the second value of independent variable of interest}

\item{B}{the upper bound value B to be used in the probability mechanism of observing false zeros}

\item{seed}{an optional seed number to control randomness}
}
\value{
a list containing:
\itemize{
\item results_parameters: a data frame for the results of model parameters,
\item results_effects: a data frame for the results of estimated mediation effect,
\item BIC: a numeric value for the BIC of the final mediation model,
\item AIC: a numeric value for the AIC of the final mediation model,
\item selected_distM: a character value for the distribution of M selected in the final mediation model
}
}
\description{
Mediation Analysis for ZEro-inflated mediators
}
\details{
A novel mediation modeling approach to address zero-inflated mediators containing true zeros and false zeros.
}
\examples{
data(zip10)
MAZE(data=zip10, distM=c('zilonm', 'zinbm', 'zipm'), 
X='X', M='Mobs', Y='Y', x1=0, x2=1, B=20, seed=1)
}
\author{
Zhigang Li \href{mailto:zhigang.li@ufl.edu}{zhigang.li@ufl.edu}
Meilin Jiang \href{mailto:meilin.jiang@ufl.edu}{meilin.jiang@ufl.edu}
}
